+++ app/models/user/querying.rb
@@ -165,6 +165,9 @@ module User::Querying
     }
     defaults[:type] = Stream::Base::TYPES_OF_POST_IN_STREAM if klass == Post
     opts = defaults.merge(opts)
+    if opts[:limit] == :all
+      opts.delete(:limit)
+    end
 
     opts[:order_field] = opts[:order].split.first.to_sym
     opts[:order_with_table] = klass.table_name + '.' + opts[:order]