+++ spec/controllers/photos_controller_spec.rb
@@ -94,6 +94,18 @@ describe PhotosController, :type => :controller do
       expect(assigns[:posts]).to eq([@bobs_photo])
     end
 
+    it "displays the correct number of photos" do
+      16.times do |i|
+        eve.post(:photo, :user_file => uploaded_photo, :to => eve.aspects.first.id, :public => true)
+      end
+      get :index, :person_id => eve.person.to_param
+      expect(response.body).to include '"photos":{"count":16}'
+
+      eve.post(:photo, :user_file => uploaded_photo, :to => eve.aspects.first.id, :public => false)
+      get :index, :person_id => eve.person.to_param
+      expect(response.body).to include '"photos":{"count":16}' # eve is not sharing with alice
+    end
+
     it "returns json when requested" do
       request.env['HTTP_ACCEPT'] = 'application/json'
       get :index, :person_id => alice.person.guid.to_s